fix(reviews): sort queue when building session
instead of sorting inside the review session component, sort when initiatizing the component from the main component

closes #823
